+++ src/bda/plone/orders/browser/views.py
@@ -704,6 +704,8 @@ class OrderViewBase(BrowserView):
         ret = list()
         for booking in self.order_data.bookings:
             obj = get_object_by_uid(self.context, booking.attrs['buyable_uid'])
+            state = vocabs.state_vocab()[booking.attrs.get('state')]
+            salaried = vocabs.salaried_vocab()[booking.attrs.get('salaried')]
             ret.append({
                 'title': booking.attrs['title'],
                 'url': obj.absolute_url(),
@@ -715,6 +717,8 @@ class OrderViewBase(BrowserView):
                 'comment': booking.attrs['buyable_comment'],
                 'quantity_unit': booking.attrs.get('quantity_unit'),
                 'currency': booking.attrs.get('currency'),
+                'state': state,
+                'salaried': salaried,
             })
         return ret